refactor(ffi): type the protocol, shrink the public surface, finish the MapSession seam
Protocol (src/protocol/):
- CreateSessionQuery and NativeTextureHandles become sealed per-backend
  subtypes (OpenGl*/Vulkan*), so the session path no longer carries 20 non-null
  assertions on handles the texture shim may or may not have sent; the reply is
  validated once, in NativeTextureHandles.fromMap.
- Typed replies: GeoPoint, ScreenPoint, FeatureHit and BoundsSpec replace
  positional lists and string-keyed maps (10 indexed accesses gone). The only
  remaining raw payload is the batch projection, which stays a Float64List on
  purpose.
- EngineEvent splits into SessionEvent and EngineScopedEvent, so the three
  offline events stop passing a fake session id of 0 and the platform adapter
  filters by type instead of by sentinel.
- protocol.dart splits into six part files, one per domain, and keeps only the
  base types and the shapes they share.

Package surface and seams:
- lib/maplibre_gl_native.dart is now the library doc plus two exports; the two
  public facades live in src/api/. The message protocol (99 declarations),
  EngineHost and the platform adapter are no longer exported: an app opts into
  the backend and keeps using maplibre_gl's own API.
- MapSession reaches the last two holders of a loose (host, sessionId) pair,
  the map view and the gesture handler, so nothing threads two values or two
  closures anymore.
- MapGestureHandler no longer depends on MapLibreFfiPlatform: it takes the
  session, the gesture flags, feature hit-testing and two event sinks, which
  makes touch behavior exercisable without a platform adapter.
- A failed map initialization is reported through FlutterError.reportError and
  shown by MapViewPlaceholder, instead of leaving a blank box and a log line.
- Internal types drop the Ffi/package prefix now that the prefix marks what is
  public: EngineCore, MapView, OrnamentConfig, GestureConfig, NativeBridge.

Docs: the frame loop and threading sections are rewritten for a reader who has
never met vsync pacing, and the state diagram becomes a two-node flowchart
(the previous one rendered as unreadable spaghetti).
