+++ maplibre_gl_native/lib/src/presentation/map/map_view.dart
@@ -6,10 +6,12 @@ import 'package:maplibre_gl_platform_interface/maplibre_gl_platform_interface.da
 
 import '../../engine/render_backend.dart';
 import '../../engine/engine_host.dart';
+import '../../engine/map_session.dart';
 import '../../protocol/protocol.dart';
 import '../platform/ffi_platform.dart';
 import '../gestures/map_gestures.dart';
 import 'map_gesture_detector.dart';
+import 'map_view_placeholder.dart';
 import '../ornaments/ornaments_overlay.dart';
 import '../platform/style_string_resolver.dart';
 import '../../native_bridge.dart';
@@ -26,8 +28,8 @@ import '../../utils/projection.dart';
 /// ornaments overlay, wires the gesture callbacks to [MapGestureHandler],
 /// and talks to the engine core exclusively through the [EngineHost] message
 /// protocol.
-class FfiMapView extends StatefulWidget {
-  const FfiMapView({
+class MapView extends StatefulWidget {
+  const MapView({
     super.key,
     required this.platform,
     required this.creationParams,
@@ -39,27 +41,35 @@ class FfiMapView extends StatefulWidget {
   final OnPlatformViewCreatedCallback onViewCreated;
 
   @override
-  State<FfiMapView> createState() => _FfiMapViewState();
+  State<MapView> createState() => _MapViewState();
 }
 
-class _FfiMapViewState extends State<FfiMapView> {
-  EngineHost? _host;
-  int? _sessionId;
+class _MapViewState extends State<MapView> {
+  /// The live session, or null until [_initialize] has created it (and again
+  /// after [dispose]). Everything that talks to the engine goes through it.
+  MapSession? _session;
   int? _textureId;
 
   bool _initStarted = false;
   bool _resizing = false;
+  Object? _initError;
   Size? _appliedLogicalSize;
   double _devicePixelRatio = 1;
 
   /// All gesture recognition and arbitration lives in the handler; the view
   /// only wires the widget callbacks to it in [build].
+  ///
+  /// The handler gets exactly the four things it needs (the session, the
+  /// gesture flags, feature hit-testing, and two event sinks) instead of the
+  /// whole platform adapter, so touch behavior can be exercised without one.
   late final MapGestureHandler _gestures = MapGestureHandler(
-    platform: widget.platform,
-    host: () => _host,
-    sessionId: () => _sessionId,
+    session: () => _session,
+    config: widget.platform.gestures,
+    features: widget.platform.features,
     cameraPitch: () => _pitch,
     mounted: () => mounted,
+    onUserPan: widget.platform.notifyUserGesture,
+    onMapLongClick: widget.platform.onMapLongClickPlatform.call,
   );
 
   // Camera bearing mirrored from engine events for the compass ornament: a
@@ -96,28 +106,27 @@ class _FfiMapViewState extends State<FfiMapView> {
   @override
   void dispose() {
     widget.platform.ornaments.removeListener(_onOrnamentsChanged);
-    final host = _host;
-    final sessionId = _sessionId;
+    final session = _session;
     final textureId = _textureId;
-    _host = null;
-    _sessionId = null;
+    _session = null;
     // Unsubscribe BEFORE disposing the notifiers _onEngineEvent writes to:
     // an event delivered in between would hit a disposed ChangeNotifier.
-    host?.removeEventListener(_onEngineEvent);
+    session?.host.removeEventListener(_onEngineEvent);
     _cameraGeneration.dispose();
     _styleGeneration.dispose();
     _metersPerPixel.dispose();
     _bearing.dispose();
-    if (host != null && sessionId != null) {
-      host.send(DisposeSessionCommand(sessionId));
+    if (session != null) {
+      session.send(DisposeSessionCommand(session.id));
     }
     if (textureId != null) {
       // The engine must have detached from the EGL surface (FIFO barrier)
       // before the platform releases the texture that backs it.
-      final barrier = host?.query(const BarrierQuery()) ?? Future.value(true);
+      final barrier =
+          session?.query(const BarrierQuery()) ?? Future.value(true);
       unawaited(
         barrier.whenComplete(
-          () => MapLibreGlNativeBridge.disposeTexture(textureId: textureId),
+          () => NativeBridge.disposeTexture(textureId: textureId),
         ),
       );
     }
@@ -135,80 +144,95 @@ class _FfiMapViewState extends State<FfiMapView> {
       final backend = RenderBackendSupport.vulkan
           ? SessionBackend.vulkan
           : SessionBackend.opengl;
-      final handles = await MapLibreGlNativeBridge.createTexture(
+      final handles = await NativeBridge.createTexture(
         physicalWidth: (logicalSize.width * devicePixelRatio).round(),
         physicalHeight: (logicalSize.height * devicePixelRatio).round(),
         backend: backend,
       );
       if (!mounted) {
-        await MapLibreGlNativeBridge.disposeTexture(
+        await NativeBridge.disposeTexture(
           textureId: handles.textureId,
         );
         return;
       }
       final logicalWidth = logicalSize.width.round();
       final logicalHeight = logicalSize.height.round();
-      final sessionId = await host.query(switch (handles.backend) {
-        SessionBackend.opengl => CreateSessionQuery.opengl(
-          textureId: handles.textureId,
-          surface: handles.surface,
-          eglDisplay: handles.eglDisplay!,
-          eglConfig: handles.eglConfig!,
-          eglContext: handles.eglContext!,
+      final sessionId = await host.query(switch (handles) {
+        final OpenGlTextureHandles gl => OpenGlSessionQuery(
+          textureId: gl.textureId,
+          surface: gl.surface,
+          eglDisplay: gl.eglDisplay,
+          eglConfig: gl.eglConfig,
+          eglContext: gl.eglContext,
           logicalWidth: logicalWidth,
           logicalHeight: logicalHeight,
           scaleFactor: devicePixelRatio,
         ),
-        SessionBackend.vulkan => CreateSessionQuery.vulkan(
-          textureId: handles.textureId,
-          surface: handles.surface,
-          vkInstance: handles.vkInstance!,
-          vkPhysicalDevice: handles.vkPhysicalDevice!,
-          vkDevice: handles.vkDevice!,
-          vkQueue: handles.vkQueue!,
-          vkQueueFamilyIndex: handles.vkQueueFamilyIndex!,
-          vkGetInstanceProcAddr: handles.vkGetInstanceProcAddr!,
-          vkGetDeviceProcAddr: handles.vkGetDeviceProcAddr!,
+        final VulkanTextureHandles vk => VulkanSessionQuery(
+          textureId: vk.textureId,
+          surface: vk.surface,
+          vkInstance: vk.vkInstance,
+          vkPhysicalDevice: vk.vkPhysicalDevice,
+          vkDevice: vk.vkDevice,
+          vkQueue: vk.vkQueue,
+          vkQueueFamilyIndex: vk.vkQueueFamilyIndex,
+          vkGetInstanceProcAddr: vk.vkGetInstanceProcAddr,
+          vkGetDeviceProcAddr: vk.vkGetDeviceProcAddr,
           logicalWidth: logicalWidth,
           logicalHeight: logicalHeight,
           scaleFactor: devicePixelRatio,
         ),
       });
-      _host = host;
-      _sessionId = sessionId;
+      final session = MapSession(host, sessionId);
+      _session = session;
       _textureId = handles.textureId;
 
-      _applyCreationParams(host, sessionId);
+      _applyCreationParams(session);
       host.addEventListener(_onEngineEvent);
-      widget.platform.attach(host, sessionId);
+      widget.platform.attach(session);
 
-      MapLibreGlNativeBridge.onSurfaceDestroyed = (textureId) {
+      NativeBridge.onSurfaceDestroyed = (textureId) {
         if (textureId == handles.textureId) {
-          host.send(SurfaceLostCommand(sessionId));
+          session.send(SurfaceLostCommand(session.id));
         }
       };
-      MapLibreGlNativeBridge.onSurfaceAvailable = (textureId) {
+      NativeBridge.onSurfaceAvailable = (textureId) {
         if (textureId == handles.textureId) {
-          unawaited(_recreateSurface(host, sessionId, textureId));
+          unawaited(_recreateSurface(session, textureId));
         }
       };
 
       setState(() {});
-      widget.onViewCreated(widget.platform.hashCode);
+      // The channel backends report the id of the platform view they created;
+      // there is no platform view here, and maplibre_gl ignores the value (it
+      // only takes the callback as the "the view exists now" signal), so 0
+      // stands for "not a platform view".
+      widget.onViewCreated(0);
     } catch (error, stackTrace) {
-      debugPrint('[maplibre_gl_native] map init failed: $error\n$stackTrace');
+      // A map that cannot initialize must not fail silently: report it the way
+      // Flutter reports any framework error (so it reaches the app's error
+      // handler and crash reporting), and leave the reason on screen in debug.
+      FlutterError.reportError(
+        FlutterErrorDetails(
+          exception: error,
+          stack: stackTrace,
+          library: 'maplibre_gl_native',
+          context: ErrorDescription('while initializing the FFI map view'),
+        ),
+      );
+      if (mounted) setState(() => _initError = error);
     }
   }
 
-  void _applyCreationParams(EngineHost host, int sessionId) {
+  void _applyCreationParams(MapSession session) {
     final params = widget.creationParams;
     final initialCamera = params['initialCameraPosition'];
     if (initialCamera is Map) {
       final target = initialCamera['target'] as List?;
       if (target != null) {
-        host.send(
+        session.send(
           JumpToCommand(
-            sessionId,
+            session.id,
             CameraSpec(
               latitude: (target[0] as num).toDouble(),
               longitude: (target[1] as num).toDouble(),
@@ -222,23 +246,22 @@ class _FfiMapViewState extends State<FfiMapView> {
     }
     final styleString = params['styleString'] as String?;
     if (styleString != null && styleString.trim().isNotEmpty) {
-      unawaited(_applyInitialStyle(host, sessionId, styleString));
+      unawaited(_applyInitialStyle(session, styleString));
     }
-    host.send(RequestRenderCommand(sessionId));
+    session.send(RequestRenderCommand(session.id));
   }
 
   /// Asset and file styles are read on the root isolate before crossing to
   /// the engine, which only accepts raw JSON or http(s) URLs.
   Future<void> _applyInitialStyle(
-    EngineHost host,
-    int sessionId,
+    MapSession session,
     String styleString,
   ) async {
     try {
       final resolved = await resolveStyleString(styleString);
-      if (!mounted || _sessionId != sessionId) return;
-      host.send(SetStyleCommand(sessionId, resolved));
-      host.send(RequestRenderCommand(sessionId));
+      if (!mounted || _session?.id != session.id) return;
+      session.send(SetStyleCommand(session.id, resolved));
+      session.send(RequestRenderCommand(session.id));
     } catch (error) {
       debugPrint(
         '[maplibre_gl_native] loading style "$styleString" failed: $error',
@@ -246,25 +269,21 @@ class _FfiMapViewState extends State<FfiMapView> {
     }
   }
 
-  Future<void> _recreateSurface(
-    EngineHost host,
-    int sessionId,
-    int textureId,
-  ) async {
+  Future<void> _recreateSurface(MapSession session, int textureId) async {
     // The engine released the swapchain when SurfaceLostCommand was
     // processed; the FIFO barrier guarantees that happened before the
     // platform destroys the old VkSurfaceKHR inside recreateSurface
     // (Vulkan mandates that destruction order).
-    await host.query(const BarrierQuery());
-    final eglSurface = await MapLibreGlNativeBridge.recreateSurface(
+    await session.query(const BarrierQuery());
+    final eglSurface = await NativeBridge.recreateSurface(
       textureId: textureId,
     );
-    if (!mounted || _sessionId != sessionId) return;
-    host.send(AttachSurfaceCommand(sessionId, eglSurface: eglSurface));
+    if (!mounted || _session?.id != session.id) return;
+    session.send(AttachSurfaceCommand(session.id, eglSurface: eglSurface));
   }
 
   void _onEngineEvent(EngineEvent event) {
-    if (event.sessionId != _sessionId) return;
+    if (event is! SessionEvent || event.sessionId != _session?.id) return;
     if (event is CameraWillChangeEvent) _cameraGeneration.value++;
     if (event is StyleLoadedEvent) _styleGeneration.value++;
     final camera = switch (event) {
@@ -297,12 +316,9 @@ class _FfiMapViewState extends State<FfiMapView> {
   }
 
   void _maybeResize(Size logicalSize, double devicePixelRatio) {
-    final host = _host;
-    final sessionId = _sessionId;
+    final session = _session;
     final textureId = _textureId;
-    if (host == null || sessionId == null || textureId == null || _resizing) {
-      return;
-    }
+    if (session == null || textureId == null || _resizing) return;
     final applied = _appliedLogicalSize;
     if (applied != null &&
         (applied.width - logicalSize.width).abs() < 1 &&
@@ -314,13 +330,12 @@ class _FfiMapViewState extends State<FfiMapView> {
     _appliedLogicalSize = logicalSize;
     _devicePixelRatio = devicePixelRatio;
     unawaited(
-      _resizeSurface(host, sessionId, textureId, logicalSize, devicePixelRatio),
+      _resizeSurface(session, textureId, logicalSize, devicePixelRatio),
     );
   }
 
   Future<void> _resizeSurface(
-    EngineHost host,
-    int sessionId,
+    MapSession session,
     int textureId,
     Size logicalSize,
     double devicePixelRatio,
@@ -330,17 +345,17 @@ class _FfiMapViewState extends State<FfiMapView> {
       // the native surface first (dropping the Vulkan swapchain while the
       // renderer stays alive; the FIFO barrier orders this across isolates)
       // so the platform never destroys a surface that is still presented to.
-      host.send(SurfaceLostCommand(sessionId));
-      await host.query(const BarrierQuery());
-      final eglSurface = await MapLibreGlNativeBridge.resizeTexture(
+      session.send(SurfaceLostCommand(session.id));
+      await session.query(const BarrierQuery());
+      final eglSurface = await NativeBridge.resizeTexture(
         textureId: textureId,
         physicalWidth: (logicalSize.width * devicePixelRatio).round(),
         physicalHeight: (logicalSize.height * devicePixelRatio).round(),
       );
-      if (!mounted || _sessionId != sessionId) return;
-      host.send(
+      if (!mounted || _session?.id != session.id) return;
+      session.send(
         ResizeSessionCommand(
-          sessionId,
+          session.id,
           logicalWidth: logicalSize.width.round(),
           logicalHeight: logicalSize.height.round(),
           scaleFactor: devicePixelRatio,
@@ -366,7 +381,7 @@ class _FfiMapViewState extends State<FfiMapView> {
               logicalSize.height >= 1) {
             unawaited(_initialize(logicalSize, devicePixelRatio));
           }
-          return const ColoredBox(color: Color(0xFF111725));
+          return MapViewPlaceholder(error: _initError);
         }
 
         _maybeResize(logicalSize, devicePixelRatio);
@@ -385,7 +400,7 @@ class _FfiMapViewState extends State<FfiMapView> {
               cameraGeneration: _cameraGeneration,
               styleGeneration: _styleGeneration,
               loadAttributions: widget.platform.getAttributions,
-              openUri: MapLibreGlNativeBridge.openUri,
+              openUri: NativeBridge.openUri,
               onCompassTap: _resetBearing,
             ),
           ],
@@ -395,11 +410,10 @@ class _FfiMapViewState extends State<FfiMapView> {
   }
 
   void _resetBearing() {
-    final host = _host;
-    final sessionId = _sessionId;
-    if (host == null || sessionId == null) return;
-    host.send(
-      EaseToCommand(sessionId, const CameraSpec(bearing: 0), durationMs: 300),
+    final session = _session;
+    if (session == null) return;
+    session.send(
+      EaseToCommand(session.id, const CameraSpec(bearing: 0), durationMs: 300),
     );
   }
 }
